+++ src/CandidatesReview/AnimalCard.scss
@@ -1,4 +1,4 @@
-@import './App';
+@import '../App';
 
 .animalCard {
   background: linear-gradient(120deg, $colorBasic, $colorDark);
