+++ src/CandidatesReview/AnimalPhotos.scss
@@ -1,4 +1,4 @@
-@import './App';
+@import '../App';
 
 div.animalPhotosContainer {
     display:flex;
