+++ src/CandidatesReview/CandidatesReview.scss
@@ -1,4 +1,4 @@
-@import "App";
+@import "../App";
 
 .page {
   display: flex;
