+++ src/CandidatesReview/CardDiffViewer.scss
@@ -1,4 +1,4 @@
-@import './App';
+@import '../App';
 
 p.attentionInfo {
     color: $colorEmphasisOrange;
