+++ src/CandidatesReview/CarouselImgViewer.scss
@@ -1,4 +1,4 @@
-@import './App';
+@import '../App';
 
 .carouselImgViewer {
     //display: grid;
